Fixed majority of loading scene issues
Managed to get scenes to work with their own loading scenes.  However, still trying to get the loading scene(s) to work with the rematch button

There seems to be some sort of bug where the loading scene freezes when quitting back to the main menu fom the pause menu.  I'm still trying to figure what could cause this...